% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{adjacency matrix function}
\usage{
adjacency(x, group)
}
\arguments{
\item{x}{description x an input matrix of dimension nobs x nvars. Each row represents an observation, each column corresponds to a covariate.}

\item{group}{description groups an integer vector indicating group sizes or as a symmetric adjacency matrix, which characterizes the grouping or graph structure of the predictors in 'x'.}
}
\description{
Definition of the incidence matrix \eqn{A}: \code{A[i, j]} equals 1 when group i includes variable j.
}
