% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.gglasso.R
\name{sel.gglasso}
\alias{sel.gglasso}
\title{selection probability with 'gglasso'}
\usage{
sel.gglasso(
  x,
  y,
  group,
  ...,
  alpha = 0.1,
  psub = 0.5,
  N.lam = 15,
  K = 100,
  eps = 1e-04,
  maxit = 1e+07
)
}
\arguments{
\item{x}{description x an input matrix of dimension nobs x nvars. Each row represents an observation, each column corresponds to a covariate.}

\item{y}{description y a response variable, where 'y = 1' corresponds to the case and 'y = 0' corresponds to the control.}

\item{group}{description groups an integer vector indicating group sizes or as a symmetric adjacency matrix, which characterizes the grouping or graph structure of the predictors in 'x'.}

\item{...}{additional arguments that can be supplied to gglasso.}

\item{alpha}{the penalty mixing parameter ranges from 0 to 1, with the default value set to 0.1.}

\item{psub}{the proportion of subsamples used for resampling is denoted as psub, the default value is 0.5.}

\item{N.lam}{the number of lambda values used for resamplings is specified, with the default value set to 15.}

\item{K}{the number of resamplings is a user-defined parameter, with the default value set to 100.}

\item{eps}{the numerical computations, the tolerance for small numerical values is \eqn{1e-04}.}

\item{maxit}{the maximum number of iterations is \eqn{1e+07}.}
}
\description{
The selection probability with 'gglasso' is computed based on resamplings.
}
